+++ cota_anegamiento/core.py
@@ -1,4 +1,11 @@
 # -*- coding: utf-8 -*-
+# Complemento QGIS «Cota de Anegamiento»
+# Copyright (C) 2026 Jose Ospina
+#
+# This program is free software; you can redistribute it and/or modify it
+# under the terms of the GNU General Public License as published by the Free
+# Software Foundation; either version 2 of the License, or (at your option)
+# any later version. See the LICENSE file for details.
 """
 Núcleo de cálculo del complemento «Cota de Anegamiento».